requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.python.org/simple
asgiref==3.3.1; python_version >= '3.5'
dj-database-url==0.5.0
django-cleanup==5.1.0
django-filter==2.4.0
django-heroku==0.3.1
django==3.1.4
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
faker==5.0.1
gunicorn==20.0.4
markdown==3.3.3
pillow==8.0.1
psycopg2==2.8.6
pyjwt==1.7.1
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.4
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
text-unidecode==1.3
whitenoise==5.2.0
